Use Logarithms for Beta
This fixes Beta(a, b) becoming zero for large values of a and b.

The idea is to use logBeta in the same way that log Gamma is used. We
can then normalise resulting likelihoods by first normalising the
log-likelihoods to ranges where exponentiation may take place.

This is done by dividing all likelihoods by the maximum likelihood, in
log space this equals a substraction. The computation is done in
log-space, of course.